+++ lib/govuk_publishing_components/presenters/schema_org.rb
@@ -1,6 +1,7 @@
 require 'govuk_publishing_components/presenters/machine_readable/page'
 require 'govuk_publishing_components/presenters/machine_readable/article_schema'
 require 'govuk_publishing_components/presenters/machine_readable/creative_work_schema'
+require 'govuk_publishing_components/presenters/machine_readable/faq_page_schema'
 require 'govuk_publishing_components/presenters/machine_readable/has_part_schema'
 require 'govuk_publishing_components/presenters/machine_readable/is_part_of_schema'
 require 'govuk_publishing_components/presenters/machine_readable/news_article_schema'
@@ -19,7 +20,9 @@ module GovukPublishingComponents
       end
 
       def structured_data
-        if page.schema == :article
+        if page.schema == :faq
+          FaqPageSchema.new(page).structured_data
+        elsif page.schema == :article
           ArticleSchema.new(page).structured_data
         elsif page.schema == :news_article
           NewsArticleSchema.new(page).structured_data